diet-analysis-backend/requirements.txt
flask
flask-cors
google-generativeai
google-api-core
json5
python-dotenv
